Add files via upload
This notebook uses linear regression algorithms to predict housing prices in California. This is the first commit and all the directories that are on my personal machine have been replaced with a placeholder, I will also upload the data files necessary to run the notebook. Also this is a current work in progress and I have completed writing code from importing the initial "csv" file to create a test set and running a few test to see what attributes are going to best to use for the ML algorithm. All cells are labeled with concise notes and I should have the finished notebook very soon.